+++ client/src/components/layouts/Header.tsx
@@ -1,13 +1,11 @@
 import { styled } from 'styled-components';
 import { Link, useNavigate } from 'react-router-dom';
 import logo from '../../assets/img/logo.png';
-import cart from '../../assets/img/cart.png';
-import user from '../../assets/img/user.png';
 // import { SearchBar } from '../../feature/SearchBar';
+import { useSetAtom } from 'jotai';
 import { menuAtom } from 'jotai/atom';
-import { isShopAtom } from 'jotai/atom';
-import { useSetAtom, useAtom } from 'jotai';
-import { AccessTokenAtom } from 'jotai/atom';
+import { AfterLogin } from './AfterLogin';
+import { BeforeLogin } from './BeforeLogin';
 
 const StyledHeaderContainer = styled.nav`
   display: flex;
@@ -16,11 +14,7 @@ const StyledHeaderContainer = styled.nav`
 const StyledLogo = styled.img`
   margin: 0.625rem;
 `;
-const StyledLoginLogo = styled.img`
-  margin: 0.625rem;
-  width: 2.3rem;
-  height: 2.3rem;
-`;
+
 const StyledChoicePage = styled.div`
   margin: 0.625rem;
   display: flex;
@@ -43,22 +37,7 @@ const StyledNavContainer = styled.div`
   justify-content: flex-end;
   flex: 1;
 `;
-const StyledButton = styled.button`
-  background-color: var(--green-100);
-  margin-right: 10px;
-  width: 4rem;
-  height: 2.5rem;
-  border: 1px solid var(--white);
-  color: var(--white);
-  &:hover {
-    cursor: pointer;
-    font-weight: bold;
-  }
 
-  &.focused {
-    font-weight: bold;
-  }
-`;
 const StyledSpan = styled.span`
   &:hover {
     cursor: pointer;
@@ -69,67 +48,23 @@ const StyledSpan = styled.span`
     font-weight: bold;
   }
 `;
-const StyledPadding = styled.div`
-  padding-right: 7rem;
-`;
-
-const StyledIconLink = styled(Link)`
-  text-decoration: none;
-`;
-
-const BeforeLogin = () => {
-  const navigation = useNavigate();
-  const handleLogin = () => {
-    navigation('/login');
-  };
-  const handlesignup = () => {
-    navigation('/signup');
-  };
-  return (
-    <>
-      <StyledPadding />
-      <StyledButton onClick={handleLogin}>로그인</StyledButton>
-      <StyledButton onClick={handlesignup}>회원가입</StyledButton>
-    </>
-  );
-};
-
-const AfterLogin = () => {
-  const navigation = useNavigate();
-
-  const handlelogout = () => {
-    navigation('/login');
-  };
-  return (
-    <>
-      <StyledPadding />
-      <StyledIconLink to={'product/like'}>
-        <StyledLoginLogo src={cart}></StyledLoginLogo>
-      </StyledIconLink>
-      <StyledIconLink to={'mypage'}>
-        <StyledLoginLogo src={user}></StyledLoginLogo>
-      </StyledIconLink>
-
-      <StyledButton onClick={handlelogout}>로그아웃</StyledButton>
-    </>
-  );
-};
 
 export const Header = () => {
   const setCurrentMenu = useSetAtom(menuAtom);
   const accessToken = localStorage.getItem('accessToken');
-  console.log('토큰:', accessToken);
   return (
     <div>
       <StyledHeaderContainer>
         <StyledLogo src={logo}></StyledLogo>
         <StyledChoicePage>
           <StyledGreen to={'/'}>
-            <StyledSpan onClick={() => setIsShop(true)}>그린</StyledSpan>
+            <StyledSpan onClick={() => setCurrentMenu('')}>그린</StyledSpan>
           </StyledGreen>
           ㅣ{/* 그린 라우팅 주소 입력*/}
           <StyledNare to={'/challenge'}>
-            <StyledSpan onClick={() => setIsShop(false)}>나래</StyledSpan>
+            <StyledSpan onClick={() => setCurrentMenu('챌린지')}>
+              나래
+            </StyledSpan>
           </StyledNare>
           {/* 나래 라우팅 주소 입력*/}
         </StyledChoicePage>
